Modernize packaging and continuous integration (#82)
* chore: modernize packaging and continuous integration

Moves package metadata from setup.py to pyproject.toml, sets Python 3.10
as the minimum supported version with NumPy 1.21 and Pandas 1.5 as the
tested dependency floors, and declares tqdm as a dependency.

Imports PyLogit only when the PyLogit-format estimation path is used and
moves it to an optional extra, so the package imports on current Python
versions. Replaces the four Travis-era workflows with a single CI workflow
that tests the minimum and latest dependency versions, checks code
quality, and validates the built distributions.

Updates tests for current Pandas releases and for the spawn multiprocessing
start method, and removes unused imports and variables flagged by the new
quality check.

* test: use fixed reference results and test on macOS and Windows

Replaces the optional urbansim.urbanchoice comparisons, which returned
without asserting when UrbanSim was absent, with seeded inputs and fixed
reference values generated from UrbanSim 3.2. Adds macOS and Windows
cells for the current-dependency job, restores UDST as the package
author, and corrects the PyLogit import comment.

* Update docs and contributor guide for the new packaging; add a docs CI job

- docs, CONTRIBUTING: replace references to the removed setup.py (install
  with `pip install -e .`, build with `python -m build`, version now lives in
  two places), drop the Python 2.7-3.8 support statement in favor of the
  3.10+ requirement, fix the dead git:// install URL, and point branch
  references at dev
- pyproject.toml: drop the `pylogit` extra, since PyLogit's current release
  does not import on any supported Python; the deferred import in mnl.py now
  raises an ImportError that points to issue #79
- MANIFEST.in: include the tests directory with its data files, so the
  test suite shipped in the sdist can actually run
- ci.yml: add a docs job that builds the Sphinx documentation with warnings
  as errors; align the build job to Python 3.14 like the other jobs
- docs/source/conf.py: `language = 'en'`, the one warning that failed the
  strict build; add a `docs` extra and update docs/README.md to match
- README: CI badge and Python 3.10 note; tests/README.md: run from the
  repository root; CHANGELOG: reword the workflow and PyLogit entries

Verified locally: 32 tests pass at the floors (Python 3.10, NumPy 1.21.6,
pandas 1.5.3) and at latest (Python 3.14, NumPy 2.5.2, pandas 3.0.5); ruff
clean; sdist and wheel build and pass `twine check --strict`; Sphinx builds
with -W on Python 3.14.

---------

+++ tests/test_mnl_urbansim.py
@@ -169,7 +169,8 @@ def test_alternative_specific_coeffs(num_alts):
     income_df = pd.DataFrame(
         np.tile(template, (num_choosers, 1)),
         columns=[
-            'boat:income', 'charter:income', 'pier:income'])
+            'boat:income', 'charter:income', 'pier:income'],
+        dtype=float)
 
     for idx, row in fish.iterrows():
         income_df.loc[idx] = income_df.loc[idx] * row['income']
@@ -186,12 +187,13 @@ def test_alternative_specific_coeffs(num_alts):
     income_df = pd.DataFrame(
         np.tile(template, (num_choosers, 1)),
         columns=[
-            'boat:income', 'charter:income', 'pier:income'])
+            'boat:income', 'charter:income', 'pier:income'],
+        dtype=float)
 
     for idx, row in fish_choosers.iterrows():
         income_df.loc[idx] = income_df.loc[idx] * row['income']
 
-    choosers_dm = pd.concat([intercept_df, income_df], axis=1)
+    choosers_dm = pd.concat([intercept_df, income_df], axis=1)  # noqa: F841
 
     # test estimation
     expected = pd.Series([
